applications apis now auto check logged in account

+++ api/routers/application.py
@@ -1,9 +1,11 @@
 from fastapi import APIRouter, Depends, HTTPException, Request, status
+from sniffio import current_async_library_cvar
 from .auth import authenticator
 from models.application import (
     ApplicationIn,
     ApplicationList,
     ApplicationOut,
+    ApplicationUpdate,
 )
 from queries.application import ApplicationQueries
 
@@ -40,30 +42,38 @@ def create_adoption_application(
 
 
 @router.get(
-    "/api/accounts/{account_id}/applications/",
+    "/api/accounts/profile/applications/",
     response_model=ApplicationList,
     summary="List all Application for an account profile page",
-    description="This lists all the applicaitons by account_id",
+    description="This lists all the applicaitons by current loggin user's account_id",
 )
 def list_account_applications(
-    account_id: str,
+    request: Request,
+    account: dict = Depends(authenticator.get_current_account_data),
     queries: ApplicationQueries = Depends(),
 ):
-    return ApplicationList(
-        applications=queries.list_account_applications(account_id)
-    )
+    if account and authenticator.cookie_name in request.cookies:
+        account_id = account["id"]
+        return ApplicationList(
+            applications=queries.list_account_applications(account_id)
+        )
+    raise not_authorized
 
 
 @router.get(
-    "/api/{rescue_id}/applications/",
+    "/api/manage/applications/",
     response_model=ApplicationList,
     summary="List all Application for Rescue",
-    description="This lists all the applicaitons by rescue_id",
+    description="This lists all the applicaitons by rescue_id of current admin/staff",
 )
 def list_adoption_applications(
-    rescue_id: str,
+    request: Request,
+    account: dict = Depends(authenticator.get_current_account_data),
     queries: ApplicationQueries = Depends(),
 ):
+    if "staff" not in account["roles"] or "admin" not in account["roles"]:
+        raise not_authorized
+    rescue_id = account["rescue_id"]
     return ApplicationList(
         applications=queries.list_applications_by_rescue_id(rescue_id)
     )
@@ -76,7 +86,8 @@ def list_adoption_applications(
     description="get application detail by application_id",
 )
 def detail_application(
-    application_id: str, queries: ApplicationQueries = Depends()
+    application_id: str,
+    queries: ApplicationQueries = Depends(),
 ):
     response = queries.detail_application(application_id)
     if response:
@@ -94,8 +105,16 @@ def detail_application(
     description="Approve an application by application_id, will check if there is an approved application based on the same pet_id, if not, change current status to Approved and change all other application have the same pet_id to Rejected",
 )
 def approve_application(
-    application_id: str, queries: ApplicationQueries = Depends()
+    application_id: str,
+    queries: ApplicationQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data),
 ):
+    if "staff" not in account["roles"] or "admin" not in account["roles"]:
+        raise not_authorized
+    elif not queries.current_staff_rescue_id_match_application(
+        application_id, account["rescue_id"]
+    ):
+        raise not_authorized
     response = queries.approve_application(application_id)
     if response:
         return response
@@ -110,8 +129,16 @@ def approve_application(
     description="Reject an application by application_id",
 )
 def reject_application(
-    application_id: str, queries: ApplicationQueries = Depends()
+    application_id: str,
+    queries: ApplicationQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data),
 ):
+    if "staff" not in account["roles"] or "admin" not in account["roles"]:
+        raise not_authorized
+    elif not queries.current_staff_rescue_id_match_application(
+        application_id, account["rescue_id"]
+    ):
+        raise not_authorized
     response = queries.reject_application(application_id)
     return response
 
@@ -119,12 +146,21 @@ def reject_application(
 @router.delete(
     "/api/applications/{application_id}/",
     summary="Delete an Application",
-    description="Delete an application by application_id, allow user to cancle an applicaiton from account profile page",
+    description="Delete an application by application_id, allow user to cancle an applicaiton from account profile page, this api will make sure current user is the owner of that applicaiton",
 )
 def delete_application(
-    application_id: str, queries: ApplicationQueries = Depends()
+    application_id: str,
+    queries: ApplicationQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data),
 ):
-    return queries.delete_application(application_id)
+    if "base" not in account["roles"]:
+        raise not_authorized
+    elif not queries.current_account_id_match_application(
+        application_id, account["id"]
+    ):
+        raise not_authorized
+    else:
+        return queries.delete_application(application_id)
 
 
 @router.patch(
@@ -135,9 +171,16 @@ def delete_application(
 )
 def update_application(
     application_id: str,
-    data: ApplicationIn,
+    data: ApplicationUpdate,
     queries: ApplicationQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data),
 ):
+    if "base" not in account["roles"]:
+        raise not_authorized
+    elif not queries.current_account_id_match_application(
+        application_id, account["id"]
+    ):
+        raise not_authorized
     response = queries.update_application(application_id, data)
     if response:
         return response
